# docs/swagger.yaml
basePath: /
definitions:
  apierror.Error:
    properties:
      error:
        type: string
    type: object
...
  battlelog.CreateTournamentInput:
    properties:
      name:
        type: string
      player_id:
        type: string
    required:
    - name
    type: object
...
  battlelog.TournamentOutput:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      player_id:
        type: string
      updated_at:
        type: string
    type: object
...
  battlelog.UpdateTournamentInput:
    properties:
      name:
        type: string
      player_id:
        type: string
    required:
    - name
    type: object
...
  user.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  user.RegisterInput:
    properties:
      default_faction:
        type: string
      email:
        type: string
      password:
        minLength: 8
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  user.UpdateUserInput:
    properties:
      default_faction:
        type: string
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - username
    type: object
  user.UserOutput:
    properties:
      created_at:
        type: string
      default_faction:
        type: string
      email:
        type: string
      id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Tournament Records and Metrics Assistant
  title: TRAMA API
  version: "1.0"
paths:
...
  /api/v1/tournaments:
    get:
      description: List all tournaments
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/battlelog.TournamentOutput'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.Error'
      summary: List tournaments
      tags:
      - tournaments
    post:
      consumes:
      - application/json
      description: Create a new tournament
      parameters:
      - description: Tournament data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/battlelog.CreateTournamentInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/battlelog.TournamentOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.Error'
      summary: Create tournament
      tags:
      - tournaments
  /api/v1/tournaments/{id}:
    delete:
      description: Delete a tournament by ID
      parameters:
      - description: Tournament ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.Error'
      summary: Delete tournament
      tags:
      - tournaments
    get:
      description: Get a tournament by ID
      parameters:
      - description: Tournament ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/battlelog.TournamentOutput'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.Error'
      summary: Get tournament
      tags:
      - tournaments
    put:
      consumes:
      - application/json
      description: Update a tournament by ID
      parameters:
      - description: Tournament ID
        in: path
        name: id
        required: true
        type: string
      - description: Tournament data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/battlelog.UpdateTournamentInput'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.Error'
      summary: Update tournament
      tags:
      - tournaments
  /api/v1/users/{id}:
    delete:
      description: Delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.Error'
      summary: Delete user
      tags:
      - users
    get:
      description: Get a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserOutput'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.Error'
      summary: Get user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserInput'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.Error'
      summary: Update user
      tags:
      - users
  /api/v1/users/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.Error'
      summary: Login
      tags:
      - users
  /api/v1/users/register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: User data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.UserOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.Error'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/apierror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.Error'
      summary: Register user
      tags:
      - users
...
swagger: "2.0"